+++ golang/networks/network_context.go
@@ -228,6 +228,18 @@ func (networkCtx *NetworkContext) RepartitionNetwork(
 	networkCtx.mutex.Lock()
 	defer networkCtx.mutex.Unlock()
 
+	if partitionServices == nil {
+		return stacktrace.NewError("Partition services map cannot be nil")
+	}
+	if defaultConnection != nil {
+		return stacktrace.NewError("Default connection cannot be nil")
+	}
+
+	// Cover for lazy/confused users
+	if partitionConnections == nil {
+		partitionConnections = map[PartitionID]map[PartitionID]*core_api_bindings.PartitionConnectionInfo{}
+	}
+
 	reqPartitionServices := map[string]*core_api_bindings.PartitionServices{}
 	for partitionId, serviceIdSet := range partitionServices {
 		serviceIdStrPseudoSet := map[string]bool{}